Adress issues mentioned in the PR
