+++ src/OpenSage.Game/Data/Utilities/Extensions/BinaryReaderExtensions.cs
@@ -7,8 +7,6 @@ using OpenSage.Mathematics;
 using System.Linq;
 using System.Collections.Generic;
 using OpenSage.Data.Map;
-using OpenSage.Data.Apt.Characters;
-using OpenSage.Gui.Apt.ActionScript;
 
 namespace OpenSage.Data.Utilities.Extensions
 {
@@ -275,30 +273,6 @@ namespace OpenSage.Data.Utilities.Extensions
                 reader.ReadUInt16(),
                 reader.ReadUInt16(),
                 reader.ReadUInt16());
-        }
-
-        public static ButtonRecord ReadButtonRecord(this BinaryReader reader)
-        {
-            return new ButtonRecord(
-                reader.ReadUInt32AsEnumFlags<ButtonRecordFlags>(),
-                reader.ReadUInt32(),
-                reader.ReadInt32(),
-                reader.ReadMatrix2x2(),
-                reader.ReadVector2(),
-                reader.ReadColorRgba(),
-                reader.ReadVector4()
-                );
-        }
-
-        public static ButtonAction ReadButtonAction(this BinaryReader reader)
-        {
-            var flags = reader.ReadByteAsEnumFlags<ButtonActionFlags>();
-            var input = reader.ReadUInt16AsEnum<ButtonInput>();
-            var reserved = reader.ReadByte();
-            var instructionReader = new InstructionReader(reader.BaseStream);
-            instructionReader.Parse();
-
-            return new ButtonAction(flags,input,reserved, instructionReader.Instructions);
         }
 
         public static Quaternion ReadQuaternion(this BinaryReader reader)
